Add files via upload
Added an update to the website, that fixes the "Pay Now" button and makes it connect to the Phantom network. Also added interactive front-end, a login page with a form submission, loading screen, card effects for each <a> element, modals that take parameters "Product Name", "Description" and "Image" in order to add products to the grid.